Fix GPU log stack corruption
It seems that the MicroProfileThreadLogGpu structure is not initialized to zero when allocated, which can result in nStackScope not being zero at the beginning of the program. I added a line to do this manually in the reset function.
This might just be me misusing the library, and I'm not sure if my patch here is the proper way to do it, but it fixed all the issues I had.