build: use system stb when possible
This patch allows microprofile to automatically use system's libstb
when available. This is particularly useful in Linux distributions,
where libraries are easily discoverable with the use of pkg-config
and similar, and also for packaging, since distributions prefer to
use shared libraries when possible.

I had to add an #if defined MICROPROFILE_SYSTEM_STB in microprofile.cpp
because when linking against a precompiled stb the
STB_SPRINTF_IMPLEMENTATION macro doesn't have to be defined and also to
avoid header clashes, because #include "stb/stb_sprintf.h" would use the
headers from the submodule even if the system one is wanted.