build: add Meson support
I was trying to package microprofile for Debian, and I had to add a
pkg-config file generator to the CMake script so that the library would
have been easily discoverable. After a bit of trials, I gave up and
decided to write a Meson script instead, since it has a built-in
pkg-config generator.

This build file can (obviously) compile the simple library just like the
CMake one, but it also provides a rich install target that is really
useful to downstream Linux distributions, and makes it also easier for
Meson users to use microprofile as a subproject.

If you are concerned with maintainability, don't worry. As you can
see there's nothing too complex, and unless you change the layout of the
.cpp and .h files, this file should not need any modification. In any
case, if some issues arise, you can always @ ping me :)

The reason why I'm upstreaming this instead of just maintaining it as a
patch in the Debian package is to simplify packaging and reduce
fragmentation - if every distro packages this library differently, its
users will have to account for x different ways of accessing the system
library, and it is not great :/